refactor out TaskStatus and SiliconCompilerError to avoid cyclical imports
